+++ sdk/python/pulumi_azurerm/batch/v20170101/batch_account.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from ... import _utilities, _tables
 from . import outputs
 from ._inputs import *
@@ -113,7 +113,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="activeJobAndJobScheduleQuota")
-    def active_job_and_job_schedule_quota(self) -> pulumi.Output[float]:
+    def active_job_and_job_schedule_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "active_job_and_job_schedule_quota")
 
     @property
@@ -126,7 +126,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="coreQuota")
-    def core_quota(self) -> pulumi.Output[float]:
+    def core_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "core_quota")
 
     @property
@@ -163,7 +163,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="poolQuota")
-    def pool_quota(self) -> pulumi.Output[float]:
+    def pool_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "pool_quota")
 
     @property
