+++ sdk/python/pulumi_azurerm/batch/v20181201/batch_account.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from ... import _utilities, _tables
 from . import outputs
 from ._inputs import *
@@ -114,7 +114,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="activeJobAndJobScheduleQuota")
-    def active_job_and_job_schedule_quota(self) -> pulumi.Output[float]:
+    def active_job_and_job_schedule_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "active_job_and_job_schedule_quota")
 
     @property
@@ -127,7 +127,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="dedicatedCoreQuota")
-    def dedicated_core_quota(self) -> pulumi.Output[float]:
+    def dedicated_core_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "dedicated_core_quota")
 
     @property
@@ -148,7 +148,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="lowPriorityCoreQuota")
-    def low_priority_core_quota(self) -> pulumi.Output[float]:
+    def low_priority_core_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "low_priority_core_quota")
 
     @property
@@ -169,7 +169,7 @@ class BatchAccount(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="poolQuota")
-    def pool_quota(self) -> pulumi.Output[float]:
+    def pool_quota(self) -> pulumi.Output[int]:
         return pulumi.get(self, "pool_quota")
 
     @property
