+++ sdk/python/pulumi_azurerm/cache/v20150801/_inputs.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from ... import _utilities, _tables
 
 __all__ = [
@@ -15,12 +15,12 @@ __all__ = [
 @pulumi.input_type
 class SkuArgs:
     def __init__(__self__, *,
-                 capacity: pulumi.Input[float],
+                 capacity: pulumi.Input[int],
                  family: pulumi.Input[str],
                  name: pulumi.Input[str]):
         """
         SKU parameters supplied to the create Redis operation.
-        :param pulumi.Input[float] capacity: What size of Redis cache to deploy. Valid values: for C family (0, 1, 2, 3, 4, 5, 6), for P family (1, 2, 3, 4).
+        :param pulumi.Input[int] capacity: What size of Redis cache to deploy. Valid values: for C family (0, 1, 2, 3, 4, 5, 6), for P family (1, 2, 3, 4).
         :param pulumi.Input[str] family: Which family to use. Valid values: (C, P).
         :param pulumi.Input[str] name: What type of Redis cache to deploy. Valid values: (Basic, Standard, Premium).
         """
@@ -30,14 +30,14 @@ class SkuArgs:
 
     @property
     @pulumi.getter
-    def capacity(self) -> pulumi.Input[float]:
+    def capacity(self) -> pulumi.Input[int]:
         """
         What size of Redis cache to deploy. Valid values: for C family (0, 1, 2, 3, 4, 5, 6), for P family (1, 2, 3, 4).
         """
         return pulumi.get(self, "capacity")
 
     @capacity.setter
-    def capacity(self, value: pulumi.Input[float]):
+    def capacity(self, value: pulumi.Input[int]):
         pulumi.set(self, "capacity", value)
 
     @property
