+++ sdk/python/pulumi_azurerm/cache/v20171001/outputs.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from ... import _utilities, _tables
 
 __all__ = [
@@ -83,12 +83,12 @@ class ScheduleEntryResponse(dict):
     """
     def __init__(__self__, *,
                  day_of_week: str,
-                 start_hour_utc: float,
+                 start_hour_utc: int,
                  maintenance_window: Optional[str] = None):
         """
         Patch schedule entry for a Premium Redis Cache.
         :param str day_of_week: Day of the week when a cache can be patched.
-        :param float start_hour_utc: Start hour after which cache patching can start.
+        :param int start_hour_utc: Start hour after which cache patching can start.
         :param str maintenance_window: ISO8601 timespan specifying how much time cache patching can take. 
         """
         pulumi.set(__self__, "day_of_week", day_of_week)
@@ -106,7 +106,7 @@ class ScheduleEntryResponse(dict):
 
     @property
     @pulumi.getter(name="startHourUtc")
-    def start_hour_utc(self) -> float:
+    def start_hour_utc(self) -> int:
         """
         Start hour after which cache patching can start.
         """
@@ -130,12 +130,12 @@ class SkuResponse(dict):
     SKU parameters supplied to the create Redis operation.
     """
     def __init__(__self__, *,
-                 capacity: float,
+                 capacity: int,
                  family: str,
                  name: str):
         """
         SKU parameters supplied to the create Redis operation.
-        :param float capacity: The size of the Redis cache to deploy. Valid values: for C (Basic/Standard) family (0, 1, 2, 3, 4, 5, 6), for P (Premium) family (1, 2, 3, 4).
+        :param int capacity: The size of the Redis cache to deploy. Valid values: for C (Basic/Standard) family (0, 1, 2, 3, 4, 5, 6), for P (Premium) family (1, 2, 3, 4).
         :param str family: The SKU family to use. Valid values: (C, P). (C = Basic/Standard, P = Premium).
         :param str name: The type of Redis cache to deploy. Valid values: (Basic, Standard, Premium)
         """
@@ -145,7 +145,7 @@ class SkuResponse(dict):
 
     @property
     @pulumi.getter
-    def capacity(self) -> float:
+    def capacity(self) -> int:
         """
         The size of the Redis cache to deploy. Valid values: for C (Basic/Standard) family (0, 1, 2, 3, 4, 5, 6), for P (Premium) family (1, 2, 3, 4).
         """
